requirements.txt
# Web Frameworks
streamlit==1.29.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==21.2.0

# HTTP & API
requests==2.31.0
httpx==0.27.0

# AI & LLM
openai>=1.12.0,<2.0.0
langchain>=0.1.20,<1.0.0
langchain-openai>=0.1.0,<1.0.0
langchain-groq>=0.1.0

# Data & Config
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary==2.9.11

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Auth & OAuth
authlib==1.3.0
httpx==0.27.0
email-validator==2.1.0
itsdangerous==2.1.2